Removed sources of compiler warnings (which are now treated as errors):
Used USE_PARAMETER for unused function arguments.
Removed declared/set but not used variables.
Put #if defined ... #endif around unused functions.
Used correct % conversion characters in printf statements for the type
of value being printed.
Had to cast pointer-to-function to void * in < and > comparisons to
avoid compiler warnings in compare_FE_field_cm_information. Perhaps it
is a little dodgy to compare pointers to functions for address?