Hiding the User_interface structure and providing access functions for all the members that are used.
Changing to using our own unix select and therefore taking control off of X.
This involves a small amount of code copied from the Xt reference
implementation which is contained in call_work_procedures.  A compile time
define USE_XTAPP_CONTEXT allows you to build a version that implements the
event dispatcher using the XtApp functions.

This should help in porting to other operating systems and allows us to use
the time_keeper and sockets in a server mode without an X display at all,
which is good for testing.

Sorted out the property notify stuff so that the plugin works more reliably
and the callbacks clean up properly.