Committed the cmgui_api_headers example. This example attempts to use SWIG to wrap whole headers from the cmgui API. So far, cmiss_node.h, cmiss_element.h, cmiss_region.h, cmiss_field_image_processing.h, cmiss_field_logical_operators.h and cmiss_field_arithmetic_operators.h are fully wrapped.
Make the modules in the cmgui_api_headers/cmiss folder. The modules are loaded as a part of the cmiss package in the python test script cmgui_api_headers_test.py. This python test script writes to file the coordinate field of the loaded mesh, and the square of the scalar field in the loaded mesh.

https://tracker.physiomeproject.org/show_bug.cgi?id=1527