Use group subelement handling mode for adding/removing objects to/from groups, rather than cmgui-specific logic. Known issue: removing lines always removes nodes used by them regardless of how many neighbouring lines are in the group; problem stems from Zinc not having faces for line elements.
Change node and element tool to clear the entire selection rather than just nodes/elements when clicking away from objects. Fix bug in element tool where the element was always added on clicking, so now it only does so if not already in the group; this allows shift-click to remove the element as expected. Fix unexpected behaviour in node tool which cleared the selection when destroying nodes; expect nodes that could not be destroyed to remain selected.

Switch to using subelement handling for selection group; this is a trial which we'll review prior to release. One known issue with the new behaviour is that selecting an element and destroying it currently destroys all its faces, including those in-use by other elements.

https://tracker.physiomeproject.org/show_bug.cgi?id=3745